przywrocenie do stranu uzywalnosci autoryzacji

+++ SOAP/scs.php
@@ -8,7 +8,8 @@ $utime_array = explode(' ', microtime());
 
 $ts=date('Ymd_Hi_s_').sprintf('%06x',(int)(($utime_array[0]*0xffffff)&0xffffff));
 
-$find_del = exec('/usr/bin/find /home/company/ftp/b2b-sklep/ups_seller2/SOAP/log/ -type f  -name "server2*" -cmin +2880 -delete -exec /bin/echo {} \; | wc -l', $output, $ret_find);
+if (strtoupper(substr(PHP_OS, 0, 3)) != 'WIN')
+	$find_del = exec('/usr/bin/find /home/company/ftp/b2b-sklep/ups_seller2/SOAP/log/ -type f -name "server2*" -cmin +2880 -delete -exec /bin/echo {} \; | wc -l', $output, $ret_find);
 
 $hdr = file_get_contents('php://input');
 $mthd_exp  = '/\:body><([a-z0-9\:]+)[\ >]/i';
